+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix dropdown field options not saving in the form builder when no default option is selected. ([#2795](https://github.com/verbb/formie/issues/2795))
 - Fix replacing a deleted field with a new field using the same handle requiring a two-step save in the form builder. ([#1854](https://github.com/verbb/formie/issues/1854))
 - Fix submit button conditions not applying on the front end. ([#2568](https://github.com/verbb/formie/issues/2568))
+- Fix checkbox fields saving their default option values when explicitly unchecked on submit. ([#2457](https://github.com/verbb/formie/issues/2457))
 
 ## 4.0.0-beta.2 - 2026-06-03